The computer will now automatically determine the ace value for the player.
The added functionality will help to add a dealer to the game. The
current implementation might cause the player to bust if they are dealt
two aces, but I believe it should work in the majority of cases.
Additionally, I cleaned up and separated some of the code into
individual methods; the separate methods should help when I attempt to
add a dealer.